Create logging infrastructure module
