+++ floris/utilities.py
@@ -12,11 +12,7 @@
  
 # See https://floris.readthedocs.io for documentation
  
-
 import numpy as np
-import coloredlogs
-import logging
-from datetime import datetime
 
 
 class Vec3():
@@ -195,93 +191,3 @@ def wrap_360(x):
     x = np.where(x < 0., x + 360., x)
     x = np.where(x >= 360., x - 360., x)
     return x
-
-
-class LogClass:
-    """
-    TODO
-    """
-
-    class __LogClass:
-        """
-        TODO
-        """
-
-        def __init__(self, param_dict):
-
-            self.log_to_console = False
-            self.console_level = 'WARNING'
-            self.log_to_file = False
-            self.file_level = 'WARNING'
-
-            # TODO: what if it IS None?
-            if param_dict is not None:
-                for key in param_dict:
-                    if key == 'console':
-                        self.log_to_console = param_dict[key]['enable']
-                        self.console_level = param_dict[key]['level']
-
-                    if key == 'file':
-                        self.log_to_file = param_dict[key]['enable']
-                        self.file_level = param_dict[key]['level']
-
-    instance = None
-
-    def __init__(self, arg):
-        if not LogClass.instance:
-            LogClass.instance = self.__LogClass(arg)
-
-    def __getattr__(self, name):
-        return getattr(self.instance, name)
-
-    def __setattr__(self, name, value):
-        self.instance.__setattr__(name, value)
-
-
-def setup_logger(name, logging_dict=None, floris=None):
-    log_class = LogClass(logging_dict)
-
-    logger = logging.getLogger(name)
-    logger.setLevel(logging.DEBUG)
-    # level = 'WARNING'
-    # level_styles = {'warning': {'color': 'red', 'bold': False}}
-    fmt_console = '%(name)s %(levelname)s %(message)s'
-    fmt_file = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-
-    file_name = 'floris_{:%Y-%m-%d-%H_%M_%S}.log'.format(datetime.now())
-
-    if log_class.log_to_console:
-        console_handler = logging.StreamHandler()
-        console_handler.setLevel(log_class.console_level)
-        console_format = coloredlogs.ColoredFormatter(
-            # level_styles=level_styles,
-            fmt=fmt_console)
-        console_handler.setFormatter(console_format)
-        console_handler.addFilter(TracebackInfoFilter(clear=True))
-        logger.addHandler(console_handler)
-
-    if log_class.log_to_file:
-        file_handler = logging.FileHandler(file_name)
-        file_handler.setLevel(log_class.file_level)
-        file_format = logging.Formatter(fmt_file)
-        file_handler.setFormatter(file_format)
-        file_handler.addFilter(TracebackInfoFilter(clear=False))
-        logger.addHandler(file_handler)
-
-    return logger
-
-
-class TracebackInfoFilter(logging.Filter):
-    """Clear or restore the exception on log records"""
-
-    def __init__(self, clear=True):
-        self.clear = clear
-
-    def filter(self, record):
-        if self.clear:
-            record._stack_info_hidden, record.stack_info = \
-                                                        record.stack_info, None
-        elif hasattr(record, "_stack_info_hidden"):
-            record.stack_info = record._stack_info_hidden
-            del record._stack_info_hidden
-        return True
